Updated Objective
-Added description
-Added walkthrough
-Updated objective.json, example, and validator to match objective direction
