+++ levels/lovelace_tower/objectives/api-05-get-patch/example.js
@@ -1,0 +1,7 @@
+async function getAndPatchCorruptedInscription(guid) {
+  // API divination endpoint: https://twilio.com/quest/magic/divination?target=lovelace_secret_statue
+  // TODO: get and patch the corrupted inscription!
+}
+
+// Replace {{RANDOM_GUID}} with any alphanumeric string (e.g. asngrue#@$55r9)
+getAndPatchCorruptedInscription("{{RANDOM_GUID}}");
