+++ levels/lovelace_tower/objectives/api-05-get-patch/walkthrough.md
@@ -1,0 +1,9 @@
+# Repairing the corrupted inscription
+
+This task asks us to use all that we've learned from the past 4 objectives. Try to recall the things we worked with and put them to use here in order to repair the corrupted inscription!
+
+## Helpful links
+
+- [MDN Fetch](https://developer.mozilla.org/en-US/docs/Web/API/Fetch_API)
+- [MDN Promises](https://developer.mozilla.org/en-US/docs/Learn/JavaScript/Asynchronous/Promises)
+- [MDN HTTP Verbs](https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods)
