--- levels/lovelace_tower/objectives/api-05-get-post/description.md
@@ -1,18 +1,0 @@
-# Get and Post
-
-<div class="aside">
-<h3>Requirements</h3>
-<ul>
-  <li>Create an async function called `getAuthorAndPostBookReview`.</li>
-  <li>This function receives a number that represents a unique id.</li>
-  <li>Use `window.fetch` to `GET` json data from the "divination" endpoint.</li>
-  <li>Use the `json` method to parse the json response.</li>
-  <li>Get an author's book from the data.</li>
-  <li>Use `window.fetch` to `POST` a review with the `id` argument, directed at the "review" endpoint for the book you chose.</li>
-  <li>Once you're done, press <em>HACK</em>.</li>
-</ul>
-</div>
-
-Content goes here...
-
-Once you've writen this function, click the _HACK_ button!
