--- levels/lovelace_tower/objectives/api-05-get-post/example.js
@@ -1,5 +1,0 @@
-async function getAuthorAndPostBookReview(id) {
-  // API book endpoint: https://twilio.com/quest/magic/divination/books
-  // API review endpoint: https://twilio.com/quest/magic/divination/reviews
-  // TODO: "GET" the divination data, choose an author's book, and "POST" a review for it!
-}
