--- levels/lovelace_tower/objectives/api-05-get-post/objective.json
@@ -1,15 +1,0 @@
-{
-  "title": "API Academy: Get and Post",
-  "description": "Write an async function that \"GET\"s data from a Web API, and \"POST\"s using data from the response.",
-  "show_ide": true,
-  "validation_fields": [
-    {
-      "name": "prompt",
-      "type": "prompt",
-      "label": "Write a function called \"getAuthorAndPostBookReview\" that makes a \"GET\" request to the \"divination\" Web API, chooses an author's book, and \"POST\"s a review for it."
-    }
-  ],
-  "rewards": {
-    "xp": 100
-  }
-}
