--- levels/lovelace_tower/objectives/api-05-get-post/walkthrough.md
@@ -1,3 +1,0 @@
-# Walkthrough
-
-Walkthrough goes here...
